+++ packages/shared-utils/src/index.ts
@@ -1,1 +1,5 @@
 export const PACKAGE_NAME = '@finanzas/shared-utils';
+
+export * from './money';
+export * from './dates';
+export * from './rut';